Initial upload of MK_2, the second iteration of my EngD Project ABM model
"MK_2" is the second iteration of my EngD project model. It is adapted
from the MASON demo, "Gridlock", made by Sarah Wise, Mark Coletti, and
Andrew Crooks. The model imports a number of GIS shapefiles, creates
multiple sets of agents, and moves those agents from A to B when they
are limited to paths defined by the GIS data.